pyproject.toml
[project]
name = "circle-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "notify-py>=0.3.43",
    "psutil>=7.1.1",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
    "typer>=0.20.0",
]


[project.scripts]
circle = "circle.main:app"
